Fixed id Long to Integer Added tests for update and delete() to User, Channel and Message RepositoriesTests

+++ src/main/java/com/example/CrmBackend/model/Message.java
@@ -16,7 +16,7 @@ public class Message {
   /** Id of user */
   @Id
   @GeneratedValue(strategy = GenerationType.IDENTITY)
-  private Long id;
+  private Integer id;
 
   /** Content of the message */
   @Column(name = "content")
@@ -49,14 +49,14 @@ public class Message {
   /**
    * @return id
    */
-  public Long getId() {
+  public Integer getId() {
     return id;
   }
 
   /**
-   * @return id
+   * @param id id of message
    */
-  public void setId(Long id) {
+  public void setId(Integer id) {
     this.id = id;
   }
 
@@ -68,14 +68,14 @@ public class Message {
   }
 
   /**
-   * @param content ccontent of message
+   * @param content content of message
    */
   public void setContent(String content) {
     this.content = content;
   }
 
   /**
-   * @param userId Id of user
+   * @return id
    */
   public Integer getUserId() {
     return userId;
